Updated file paths, added demo home page
