+++ lib/demos/eyesight_stats/eyesight_page.dart
@@ -3,8 +3,8 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:healthcare_superplatform/widgets/custom_appbar.dart';
-import 'package:healthcare_superplatform/widgets/custom_table_widget.dart';
+import 'package:healthcare_superplatform/unused/widgets/custom_appbar.dart';
+import 'package:healthcare_superplatform/demos/eyesight_stats/widgets/custom_table_widget.dart';
 
 class EyesightPage extends StatefulWidget {
   const EyesightPage({super.key});
