+++ lib/unused/pages/self_diagnose/self_diagnose_fatigue.dart
