+++ lib/unused/widgets/available_devices_list.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../models/device.dart';
+import '../../models/device.dart';
 
 class AvailableDevicesList extends StatelessWidget {
   final bool isScanning;
@@ -41,24 +41,30 @@ class AvailableDevicesList extends StatelessWidget {
         itemBuilder: (context, index) {
           final device = availableDevices[index];
           // Check if device is already connected
-          final bool isConnected = connectedDevices.any((d) => d.id == device.id);
-          
+          final bool isConnected = connectedDevices.any(
+            (d) => d.id == device.id,
+          );
+
           return Card(
             margin: const EdgeInsets.only(bottom: 8.0),
             child: ListTile(
               leading: Icon(device.icon, size: 32),
               title: Text(device.name),
               subtitle: Text('${device.manufacturer} • ${device.type}'),
-              trailing: isConnected
-                  ? const Text('Connected', style: TextStyle(color: Colors.green))
-                  : ElevatedButton(
-                      onPressed: () => onConnect(device),
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colors.blue.shade100,
-                        foregroundColor: Colors.blue.shade900,
+              trailing:
+                  isConnected
+                      ? const Text(
+                        'Connected',
+                        style: TextStyle(color: Colors.green),
+                      )
+                      : ElevatedButton(
+                        onPressed: () => onConnect(device),
+                        style: ElevatedButton.styleFrom(
+                          backgroundColor: Colors.blue.shade100,
+                          foregroundColor: Colors.blue.shade900,
+                        ),
+                        child: const Text('Connect'),
                       ),
-                      child: const Text('Connect'),
-                    ),
             ),
           );
         },
