+++ lib/unused/widgets/medication_card.dart
@@ -1,32 +1,29 @@
 import 'package:flutter/material.dart';
-import '../models/medication.dart';
+import '../../models/medication.dart';
 import 'package:intl/intl.dart';
 import '../pages/prescription_detail_page.dart';
 
 class MedicationCard extends StatelessWidget {
   final Medication medication;
 
-  const MedicationCard({
-    Key? key,
-    required this.medication,
-  }) : super(key: key);
+  const MedicationCard({Key? key, required this.medication}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final bool needsRefill = medication.refillDate.difference(DateTime.now()).inDays <= 3;
-    
+    final bool needsRefill =
+        medication.refillDate.difference(DateTime.now()).inDays <= 3;
+
     return Card(
       margin: const EdgeInsets.only(bottom: 12.0),
       elevation: 2,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(12),
-      ),
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
       child: InkWell(
         onTap: () {
           Navigator.push(
             context,
             MaterialPageRoute(
-              builder: (context) => PrescriptionDetailPage(medication: medication),
+              builder:
+                  (context) => PrescriptionDetailPage(medication: medication),
             ),
           );
         },
@@ -81,10 +78,7 @@ class MedicationCard extends StatelessWidget {
                 children: [
                   Text(
                     'Expires',
-                    style: TextStyle(
-                      fontSize: 12,
-                      color: Colors.grey.shade600,
-                    ),
+                    style: TextStyle(fontSize: 12, color: Colors.grey.shade600),
                   ),
                   Text(
                     DateFormat('MMM d, yyyy').format(medication.expirationDate),
